requirements.txt
openpyxl==3.1.2
pandas==2.2.3
flask==2.2.5
ruamel.yaml==0.17.21
pytest==8.4.0
requests==2.32.5
python-json-logger==2.0.6
prometheus_client==0.16.0
boto3==1.28.35
